% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_chart.R
\name{bar_chart}
\alias{bar_chart}
\title{Bar Chart}
\usage{
bar_chart(
  data,
  x,
  y = NULL,
  stat = "count",
  width = 0.7,
  prop = FALSE,
  top = NULL,
  fill = "#4E79A7",
  fill_var = NULL,
  facet_var = NULL,
  scales = "free",
  ncol = 3,
  position = "dodge",
  flip = FALSE,
  legend_position = "right",
  title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  highlight = NULL,
  palette_gradient = "Green-Gold",
  label_var = NULL,
  label_color = "gray20",
  label_size = 3.1,
  expand_mult = c(0.05, 0.1)
)
}
\arguments{
\item{data}{A data-frame or tibble.}

\item{x}{A categorical variable.}

\item{y}{A numeric variable when \code{stat = "count"}.}

\item{stat}{character. How to calculate bar heights. Options include "count" or "identity".}

\item{width}{numeric. Width of bars.}

\item{prop}{logical. Whether or not to display proportions instead of counts. Overrided if stat = "identity".}

\item{top}{integer. Top n levels.}

\item{fill}{character. Color of the bars. Overrided if fill_var is specified.}

\item{fill_var}{A categorical variable to group and fill.}

\item{facet_var}{A categorical variable used to create facet panels.}

\item{scales}{character. control facet x & y-axis ranges. Options include "fixed", "free", "free_y", "free_x"}

\item{ncol}{integer. Number of facet columns.}

\item{position}{character. Position adjustment. Options include "stack", fill" or "dodge".}

\item{flip}{logical. whether or not to flip the chart horizontally.}

\item{legend_position}{A character indicating legend position. Options include "right", "left", "top", "bottom", "none".}

\item{title}{character. Title for the plot.}

\item{x_lab}{character. X-axis label for the plot.}

\item{y_lab}{character. Y-axis label for the plot.}

\item{highlight}{A vector of categories to highlight. Overrided if fill_var is specified.}

\item{palette_gradient}{character. Color palette when fill_var is a numeric variable.}

\item{label_var}{The categorical or numeric variable used to label the bars.}

\item{label_color}{character. Label color.}

\item{label_size}{numeric. Label Size.}

\item{expand_mult}{character. The expansion multiplier for the y-axis.}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot.
}
\description{
Plot a Bar Chart in ggplot2
}
\examples{
bar_chart(ggplot2::diamonds, x = clarity)

bar_chart(ggplot2::diamonds, x = clarity, prop = TRUE)

bar_chart(ggplot2::diamonds, x = clarity, flip = TRUE)

bar_chart(ggplot2::diamonds, x = clarity, top = 5)

bar_chart(ggplot2::diamonds, x = clarity, fill_var = cut, flip = TRUE)

bar_chart(ggplot2::diamonds, x = clarity, fill_var = cut, flip = TRUE,
          position = "stack")

bar_chart(ggplot2::diamonds, x = clarity, fill_var = cut,
          position = "fill", flip = TRUE)

bar_chart(ggplot2::diamonds, x = clarity, fill_var = cut, position = "fill",
          flip = TRUE, facet_var = color, legend_position = "bottom")

}
