% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_chart.R
\name{bubble_chart}
\alias{bubble_chart}
\title{Bubble Chart}
\usage{
bubble_chart(
  data,
  x,
  y,
  size_var,
  fill_var,
  facet_var = NULL,
  scale_size = c(0.5, 20),
  alpha = 0.8,
  stroke = 1.5,
  trans_x = "identity",
  trans_y = "identity",
  title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  scales = "free",
  ncol = 3,
  legend_position = "right",
  interactive = FALSE
)
}
\arguments{
\item{data}{A data.frame or tibble.}

\item{x}{A numeric variable.}

\item{y}{A numeric variable.}

\item{size_var}{A numeric variable indicating the size.}

\item{fill_var}{A categorical variable to group and fill.}

\item{facet_var}{A categorical variable used to create facet panels.}

\item{scale_size}{A numeric vector with the minimum and maximum size.}

\item{alpha}{numeric. Opacity. Range: (0, 1).}

\item{stroke}{numeric. Stroke width (outline).}

\item{trans_x}{character. X-axis transformation.}

\item{trans_y}{character. Y-axis transformation.}

\item{title}{character. Title for the plot.}

\item{x_lab}{character. X-axis label for the plot.}

\item{y_lab}{character. Y-axis label for the plot.}

\item{scales}{character. Control facet x & y-axis ranges. Options include "fixed", "free", "free_y", "free_x"}

\item{ncol}{integer. Number of facet columns.}

\item{legend_position}{character. Legend position. Options include "right", "left", "top", "bottom", "none".}

\item{interactive}{logical. Whether or not to return an interactive plot (plotly).}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot
}
\description{
Plot Bubble Chart in ggplot2
}
\examples{
library(dplyr)
library(gapminder)

data <- gapminder \%>\% filter(year=="2002") \%>\%
  dplyr::select(-year)

data_bubble <- data \%>\%
  mutate(pop=pop/1000000) \%>\%
  arrange(desc(pop)) \%>\%
  mutate(country = factor(country, country))

bubble_chart(
  data = data_bubble,
  x = gdpPercap,
  y = lifeExp,
  size_var = pop,
  fill_var = continent,
  title = "GDPpercap vs Life Exp",
  x_lab = "Gdp Per Cap",
  y_lab = "Life Exp",
  trans_x = "log10",
  legend_position = "none",
  facet_var = continent
)


}
