% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_kpi_line.R
\name{card_kpi_line}
\alias{card_kpi_line}
\title{Kpi card with line chart}
\usage{
card_kpi_line(
  data,
  date,
  value,
  prior_year,
  current_year,
  summarise_by = "month",
  line_color = "#4E79A7",
  line_width = 1.2,
  area_color = "gray",
  area_alpha = 0.4,
  label_size = 3.5,
  show_label = TRUE,
  expand_mult_x = c(0.04, 0.12),
  title = NULL,
  title_size = NULL,
  title_hjust = 0.5,
  legend_position = "top",
  pct = FALSE,
  value_text = NULL,
  value_size = 10,
  value_color = "gray20",
  value_coord = c(1, 1.8),
  yoy_text = NULL,
  yoy_size = 5,
  yoy_coord = c(1, 1.71),
  yoy_color = NULL,
  rect_coord = c(0.978, 1.021, 1.67, 1.74),
  rect_color = NULL,
  rect_alpha = 0.2,
  text_top = NULL,
  text_top_size = 4,
  text_top_coord = c(1, 1.87),
  text_top_color = "gray20",
  text_bottom = NULL,
  text_bottom_size = 4,
  text_bottom_coord = c(1, 1.64),
  text_bottom_color = "gray20",
  widths = c(1, 1.5),
  show_axis_text = TRUE,
  interactive = FALSE
)
}
\arguments{
\item{data}{A data-frame or tibble.}

\item{date}{A date or date time column.}

\item{value}{A numeric column.}

\item{prior_year}{integer. Prior year.}

\item{current_year}{integer. Current year.}

\item{summarise_by}{character. Time unit to summarise.}

\item{line_color}{character. Line Color.}

\item{line_width}{numeric. Line_width.}

\item{area_color}{character. The color for the area chart (Prior year).}

\item{area_alpha}{numeric. Opacity for the area chart (Prior year).}

\item{label_size}{numeric. Label size.}

\item{show_label}{logical. Whether or not to include label.}

\item{expand_mult_x}{A numeric vector of length 2 containing the expansion multiplier.}

\item{title}{character. Title for the plot.}

\item{title_size}{numeric. The size of the title.}

\item{title_hjust}{numeric. The hjust of the title.}

\item{legend_position}{character. Legend position. Options include "right", "left", "top", "bottom", "none".}

\item{pct}{logical. Whether or not to format the values as percentages.}

\item{value_text}{character. The text for the current year.}

\item{value_size}{numeric. The text size for \code{value_text}.}

\item{value_color}{character. The text color for \code{value_text}.}

\item{value_coord}{numeric. A vector of length 2 containing the coordinates (x,y) for \code{value_text}.}

\item{yoy_text}{character. The text for YoY.}

\item{yoy_size}{numeric. The size for for \code{yoy_text}.}

\item{yoy_coord}{numeric. A vector of length 2 containing the coordinates (x,y) for \code{yoy_text}.}

\item{yoy_color}{character. The text color for \code{yoy_text}.}

\item{rect_coord}{numeric. A vector of length 4 containing the coordinates of the rectangle to be drawn around the \code{yoy_text}. The values are xmin, xmax, ymin, ymax.}

\item{rect_color}{character. The color of the rectangle to be drawn around the \code{yoy_text}.}

\item{rect_alpha}{numeric. Opacity for the rectangle to be drawn around the \code{yoy_text}.}

\item{text_top}{character. The text to display at the top of the \code{value_text}.}

\item{text_top_size}{numeric. The size of the text to display at the top of the \code{value_text}.}

\item{text_top_coord}{numeric. A vector of length 2 containing the coordinates (x,y) to display at the top of the \code{value_text}.}

\item{text_top_color}{character. The color of text to display at the top of the \code{value_text}.}

\item{text_bottom}{character. The text to display at the bottom of the \code{yoy_text}.}

\item{text_bottom_size}{numeric. The size of the text dto display at the bottom of the \code{yoy_text}.}

\item{text_bottom_coord}{numeric. A vector of length 2 containing the coordinates (x,y) the text to display at the bottom of the \code{yoy_text}.}

\item{text_bottom_color}{character. The color of text to display at the bottom of the \code{yoy_text}.}

\item{widths}{numeric. A vector of length 2 containing the relative widths of the line chart and kpi.}

\item{show_axis_text}{logical. Whether or not to show x-axis text.}

\item{interactive}{logical. Whether or not to return an interactive plot (plotly).}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot.
}
\description{
Kpi card with line chart in ggplot2
}
\examples{
library(dplyr)

data <- readr::read_csv(
  file = "https://raw.githubusercontent.com/zpio/datasets/main/sample_superstore.csv"
)

data \%>\%
  card_kpi_line(
    date = order_date,
    value = sales,
    prior_year = 2020,
    current_year = 2021
  )

data \%>\%
  filter(sub_category == "Machines") \%>\%
  card_kpi_line(
    date = order_date,
    value = sales,
    prior_year = 2020,
    current_year = 2021,
    title = "Subcategory: Machines",
    title_size = 10,
    line_color = "cyan3",
    area_color = "cyan",
    area_alpha = 0.2
  )


}
