% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate}
\alias{correlate}
\title{Correlations}
\usage{
correlate(data, method = "pearson")
}
\arguments{
\item{data}{A data frame or tibble.}

\item{method}{character. Correlation method. One of "pearson", "kendall", or "spearman".}
}
\value{
A tibble
}
\description{
This function calculate pairwise correlations between numeric columns of a data frame.
}
\examples{

correlate(iris)


}
