% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-functions.R
\name{Date_Functions}
\alias{Date_Functions}
\alias{calendarAuto}
\alias{dateAdd}
\alias{dateDiff}
\alias{dateName}
\alias{dateParse}
\alias{datePart}
\alias{dateTrunc}
\alias{year}
\alias{month}
\alias{day}
\alias{quarter}
\alias{semester}
\alias{week}
\alias{dayOfWeek}
\alias{dayOfYear}
\alias{dayOfMonth}
\alias{isWorkingDay}
\alias{isDate}
\alias{makeDate}
\alias{makeDateTime}
\alias{now}
\alias{today}
\title{Date Functions}
\usage{
calendarAuto(start_date, end_date, week_start = 1)

dateAdd(date_part, interval, date)

dateDiff(date_part, end_date, start_date)

dateName(date_part, date)

dateParse(date_format, date_string)

datePart(date_part, date)

dateTrunc(date_part, date, week_start = 1)

year(date)

month(date)

day(date)

quarter(date)

semester(date)

week(date)

dayOfWeek(date, week_start = 1)

dayOfYear(date)

dayOfMonth(date)

isWorkingDay(date, week_start = 1)

isDate(date)

makeDate(year, month, day)

makeDateTime(year, month, day, hour, min, sec, tz = "UTC")

now(tz = "")

today(tz = "")
}
\arguments{
\item{start_date}{A vector of dates.}

\item{end_date}{A vector of dates.}

\item{week_start}{character. Week start day. Default is 1, Monday.}

\item{date_part}{character. Part of the date: 'day', 'month', 'year', 'week', 'quarter', 'hour', 'minute', 'second'.}

\item{interval}{integer. Interval added to the specified date_part of that date.}

\item{date}{A vector of dates.}

\item{date_format}{character. Date format: 'ymd', 'ydm', 'mdy', etc.}

\item{date_string}{A character vector of dates.}

\item{year}{integer. Year.}

\item{month}{integer. Month.}

\item{day}{integer. Day.}

\item{hour}{integer. Hour.}

\item{min}{integer. Minute.}

\item{sec}{integer. Second.}

\item{tz}{character. Time zone. Defaults: 'UTC'.}
}
\value{
A date or date-time object the length of x
}
\description{
Perform various operations on date objects.
}
\examples{


calendarAuto("2023-01-01", "2023-01-26")

dateAdd("year", 1, as.Date("2023-01-31"))

dateDiff("months", as.Date("2023-02-28"), as.Date("2023-01-01"))

dateName('quarters', as.Date("2023-02-28"))

dateParse("ymd_hms", "2023-02-28 20:30:08.0001 pm")

datePart('month', as.Date("2023-02-28"))

dateTrunc('month', as.Date("2023-12-24"))

dateTrunc('quarter', as.Date("2023-12-24"))

dateTrunc('month', lubridate::ymd_hms("2023-01-24 10:20:20 AM"))

year(as.Date("2023-01-24"))

month(as.Date("2023-01-24"))

day(as.Date("2023-01-24"))

quarter(as.Date("2023-01-24"))

semester(as.Date("2023-01-24"))

week(as.Date("2023-01-24"))

dayOfWeek(as.Date("2023-01-24"))

dayOfYear(as.Date("2023-12-31"))

dayOfMonth(as.Date("2023-12-31"))

isWorkingDay(as.Date("2023-01-01"))

isDate(as.Date("2009-08-03"))

makeDate(2023,01,01)

makeDateTime(2023,01,01,12,50,10)

now()

today()



}
