% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptiveStats.R
\name{descriptiveStats}
\alias{descriptiveStats}
\title{Descriptive Statistic}
\usage{
descriptiveStats(data)
}
\arguments{
\item{data}{A data frame or tibble.}
}
\value{
A tibble
}
\description{
Descriptive statistics of all numeric variables from data frame.
}
\examples{
library(dplyr)

descriptiveStats(iris)

iris \%>\%
   group_by(Species) \%>\%
   descriptiveStats()


}
