% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreCategory.R
\name{exploreCategory}
\alias{exploreCategory}
\title{Explore categorical variables}
\usage{
exploreCategory(data, top = 30)
}
\arguments{
\item{data}{A data frame or tibble}

\item{top}{integer. Top n levels.}
}
\value{
A tibble
}
\description{
Explore categorical variables from data frame.
}
\examples{

exploreCategory(iris)

}
