% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreData.R
\name{exploreData}
\alias{exploreData}
\title{Explore a data frame or tibble}
\usage{
exploreData(data)
}
\arguments{
\item{data}{A data frame or tibble.}
}
\value{
A tibble.
}
\description{
Explores all variables from data frame.
}
\examples{

library(dplyr)

mtcars <- mtcars \%>\% as_tibble() \%>\%
   mutate(across(c(am, carb, cyl, gear, vs), as.factor))

iris <- as_tibble(iris)

iris \%>\%
   exploreData()

mtcars \%>\%
   exploreData()

}
