% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreNumeric.R
\name{exploreNumeric}
\alias{exploreNumeric}
\title{Explore numeric variables}
\usage{
exploreNumeric(data)
}
\arguments{
\item{data}{A data-frame or tibble}
}
\value{
A tibble
}
\description{
Explores all numeric variables from data frame.
}
\examples{
library(dplyr)

iris \%>\%
   exploreNumeric()

iris \%>\%
   select(Sepal.Length, Sepal.Width) \%>\%
   exploreNumeric()

iris \%>\%
  group_by(Species) \%>\%
  exploreNumeric()

mtcars <- mtcars \%>\% as_tibble() \%>\%
  mutate(across(c(am, carb, cyl, gear, vs), as.factor))

mtcars \%>\%
   dplyr::group_by(cyl, am) \%>\%
   exploreNumeric()

}
