% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_chart.R
\name{line_chart}
\alias{line_chart}
\title{Line Chart}
\usage{
line_chart(
  data,
  x,
  y,
  color_var = NULL,
  color = "#4E79A7",
  label_var = NULL,
  label_color = NULL,
  label_size = 3,
  label_hjust = -0.1,
  label_vjust = 0.5,
  facet_var = NULL,
  ncol = 2,
  scales = "free_y",
  title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  linewidth = 1,
  legend_position = "none",
  highlight = NULL,
  compare = FALSE,
  expand_mult_x = c(0.04, 0.2),
  check_overlap = FALSE,
  interactive = FALSE,
  show_label = TRUE
)
}
\arguments{
\item{data}{A data-frame or tibble}

\item{x}{A numeric variable}

\item{y}{A numeric variable}

\item{color_var}{A categorical variable to group and fill.}

\item{color}{character. Line color. Overrided if color_var is specified.}

\item{label_var}{A categorical variable to label.}

\item{label_color}{character. Label color.}

\item{label_size}{numeric. Size of labels.}

\item{label_hjust}{numeric. Horizontal justification of label.}

\item{label_vjust}{numeric. Vertical justification of label.}

\item{facet_var}{A categorical variable used to create facet panels.}

\item{ncol}{integer. Number of facet columns.}

\item{scales}{character. Control facet x & y-axis ranges. Options include "fixed", "free", "free_y", "free_x"}

\item{title}{character. Title for the plot.}

\item{x_lab}{character. X-axis label for the plot.}

\item{y_lab}{character. Y-axis label for the plot.}

\item{linewidth}{numeric. Line width.}

\item{legend_position}{character. Legend position. Options include "right", "left", "top", "bottom", "none".}

\item{highlight}{A vector of categories to highlight. Overrided if fill_var is specified.}

\item{compare}{logical. Whether or not to compare a group with the others.}

\item{expand_mult_x}{character. The expansion multiplier for the x-axis.}

\item{check_overlap}{logical. Whether or not to stroke the text to overlap}

\item{interactive}{logical. Whether or not to return an interactive plot (plotly).}

\item{show_label}{logical. Whether or not to display labels.}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot.
}
\description{
Plot Line Chart in ggplot2
}
\examples{

library(dplyr)

line_chart(ggplot2::economics, x = date, y = unemploy)

economics_long2 <-
  dplyr::filter(
    ggplot2::economics_long,
    variable \%in\% c("pop", "uempmed", "unemploy")
  )

line_chart(economics_long2, x = date, y = value01, color_var = variable)

line_chart(economics_long2, x = date, y = value01, color_var = variable,
           highlight = "unemploy")

line_chart(economics_long2, x = date, y = value01, color_var = variable,
           facet_var = variable)

line_chart(economics_long2, x = date, y = value01, color_var = variable,
           facet_var = variable,  scales = "fixed")

line_chart(economics_long2, x = date, y = value01, color_var = variable,
           facet_var = variable, compare = TRUE)

}
