% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBar.R
\name{plotBar}
\alias{plotBar}
\title{Explore categorical columns with bar charts}
\usage{
plotBar(
  data,
  fill_var = NULL,
  fill = NULL,
  width = 0.7,
  top = 10,
  pct = FALSE,
  position = NULL,
  ncol = 3,
  scales = "free",
  alpha = 1,
  title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  legend_position = "right",
  interactive = FALSE
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{fill_var}{A categorical column to group and fill.}

\item{fill}{character. Color of bars. Overrided if fill_var is specified.}

\item{width}{numeric. Width of the bars.}

\item{top}{integer. Top n levels.}

\item{pct}{logical. Whether or not to display proportions instead of counts. Overrided if fill_var is specified.}

\item{position}{character. Position adjustment. Options include "stack", fill" or "dodge".}

\item{ncol}{integer. Number of facet columns.}

\item{scales}{character. Control facet x & y-axis ranges. Options include "fixed", "free", "free_y", "free_x"}

\item{alpha}{numeric. Opacity. Range: (0, 1).}

\item{title}{character. Title for the plot.}

\item{x_lab}{character. X-axis label for the plot.}

\item{y_lab}{character. Y-axis label for the plot.}

\item{legend_position}{character. Legend position. Options include "right", "left", "top", "bottom", "none".}

\item{interactive}{logical. Whether or not to return an interactive plot (plotly).}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot.
}
\description{
Plot Bar Chart of all categorical columns from data frame.
}
\examples{
plotBar(ggplot2::diamonds)

plotBar(ggplot2::diamonds, fill = "steelblue")

plotBar(ggplot2::diamonds, fill = "steelblue", pct = TRUE)

plotBar(ggplot2::diamonds, fill = "steelblue", top = 5)

plotBar(ggplot2::diamonds, fill = "steelblue", top = 5, ncol = 2)

plotBar(ggplot2::diamonds, fill_var = cut)

plotBar(ggplot2::diamonds, fill_var = cut, position = "stack")

plotBar(ggplot2::diamonds, fill_var = cut, position = "dodge")

plotBar(ggplot2::diamonds, fill_var = cut, position = "fill")

}
