% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoxplot.R
\name{plotBoxplot}
\alias{plotBoxplot}
\title{Explore numeric and categorical columns with Boxplot}
\usage{
plotBoxplot(
  data,
  target,
  fill = NULL,
  ncol = 3,
  scales = "free",
  alpha = 0.7,
  outlier_shape = 19,
  outlier_color = NULL,
  outlier_size = 1.5,
  outlier_alpha = 0.7,
  width = 0.7,
  title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  flip = FALSE
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{target}{A column to group and fill.}

\item{fill}{character. Color of all boxplot. Overrided if fill_var is specified.}

\item{ncol}{integer. Number of facet columns.}

\item{scales}{character. Control facet x & y-axis ranges. Options include "fixed", "free", "free_y", "free_x".}

\item{alpha}{numeric. Opacity. Range: (0, 1).}

\item{outlier_shape}{integer. Shape of outliers.}

\item{outlier_color}{character. Color of outliers}

\item{outlier_size}{numeric. Size of outliers.}

\item{outlier_alpha}{numeric. Opacity of outliers.}

\item{width}{numeric. Width of Boxplot.}

\item{title}{character. Title for the plot.}

\item{x_lab}{character. X-axis label for the plot.}

\item{y_lab}{character. Y-axis label for the plot.}

\item{flip}{logical. Whether or not to flip cartesian coordinates.}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot.
}
\description{
Plot Boxplot of numeric columns with respect to a categorical column from data frame.
}
\examples{
plotBoxplot(iris, target = Species)

plotBoxplot(iris, target = Species, flip = TRUE)

plotBoxplot(iris, target = Species, fill = "steelblue")

mtcars <- mtcars \%>\% dplyr::as_tibble() \%>\%
  dplyr::mutate(dplyr::across(c(am, carb, cyl, gear, vs), as.factor))

plotBoxplot(mtcars, target = mpg)

plotBoxplot(mtcars, target = mpg, fill = "steelblue")

}
