% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelate.R
\name{plotCorrelate}
\alias{plotCorrelate}
\title{Explore correlations of numeric columns with bar charts.}
\usage{
plotCorrelate(
  data,
  method = "pearson",
  width = 0.8,
  label_show = TRUE,
  label_size = 3.1,
  label_color = "grey20",
  palette_gradient = "Red-Green Diverging",
  title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  expand_mult = c(0.2, 0.2),
  interactive = FALSE
)
}
\arguments{
\item{data}{A data frame or tibble}

\item{method}{character. Correlation coefficient method. Options include "pearson", "kendall" or "spearman".}

\item{width}{numeric. Width of the bars.}

\item{label_show}{logical. Whether or not to display labels.}

\item{label_size}{character. The size of the labels.}

\item{label_color}{character. The color of the labels.}

\item{palette_gradient}{character. Divergent color palette. Color palettes used in Tableau (ggthemes).}

\item{title}{character. Title for the plot.}

\item{x_lab}{character. X-axis label for the plot.}

\item{y_lab}{character. Y-axis label for the plot.}

\item{expand_mult}{A numeric vector of length 2 indicating the expansion multiplier for the x-axis.}

\item{interactive}{logical. Whether or not to return an interactive plot (plotly).}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot.
}
\description{
Explore correlations of numeric columns with bar charts.
}
\examples{

plotCorrelate(iris)

}
