% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensity.R
\name{plotDensity}
\alias{plotDensity}
\title{Explore numeric columns with Density plot}
\usage{
plotDensity(
  data,
  fill_var = NULL,
  fill = NULL,
  ncol = 3,
  bins = 30,
  alpha = 1,
  scales = "free",
  show_histogram = FALSE,
  show_boxplot = FALSE,
  adjust = 1,
  trans_x = "identity",
  trans_y = "identity",
  legend_position = "right",
  title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  interactive = FALSE
)
}
\arguments{
\item{data}{A data.frame or tibble.}

\item{fill_var}{A categorical variable to group and fill.}

\item{fill}{character. Color of all Density and Histograms. Overrided if fill_var is specified.}

\item{ncol}{integer. Number of facet columns.}

\item{bins}{integer. Number of bins. Overrided if show_histogram = FALSE.}

\item{alpha}{numeric. Opacity. Range: (0, 1).}

\item{scales}{character. Control facet x & y-axis ranges. Options include "fixed", "free", "free_y", "free_x".}

\item{show_histogram}{logical. Whether or not to include histograms. Overrided if show_boxplot = TRUE.}

\item{show_boxplot}{logical. Whether or not to include Boxplot Overrided if show_histogram = TRUE.}

\item{adjust}{numeric. Multiplicate bandwidth adjustment.}

\item{trans_x}{character. X-axis transformation.}

\item{trans_y}{character. X-axis transformation.}

\item{legend_position}{character. Legend position. Options include "right", "left", "top", "bottom", "none".}

\item{title}{character. Title for the plot.}

\item{x_lab}{character. X-axis label for the plot.}

\item{y_lab}{character. Y-axis label for the plot.}

\item{interactive}{logical. Whether or not to return an interactive plot (plotly).}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot.
}
\description{
Plot Density distribution of all numeric variables from data frame.
}
\examples{
plotDensity(iris, alpha = 0.5, ncol = 2)

plotDensity(iris, alpha = 0.5, ncol = 2, show_boxplot = TRUE)

plotDensity(iris, fill = "steelblue", alpha = 0.5, adjust = 2)

plotDensity(iris, fill = "steelblue", alpha = 0.5, adjust = 2, show_boxplot = TRUE)

plotDensity(iris, alpha = .5, show_histogram = TRUE, adjust = 2)

plotDensity(iris, fill = "steelblue", alpha = .2, show_histogram = TRUE, adjust = 2)

plotDensity(iris, alpha = .2, fill_var = Species)

plotDensity(iris, alpha = .2, fill_var = Species, legend_position = "bottom", ncol=2)

plotDensity(iris, alpha = .2, fill_var = Species, show_histogram = TRUE, adjust = 2)


}
