% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistogram.R
\name{plotHistogram}
\alias{plotHistogram}
\title{Explore numeric columns with Histogram}
\usage{
plotHistogram(
  data,
  fill_var = NULL,
  fill = NULL,
  ncol = 3,
  bins = 30,
  alpha = 1,
  scales = "free",
  show_density = FALSE,
  adjust = 1,
  trans_x = "identity",
  trans_y = "identity",
  legend_position = "right",
  title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  interactive = FALSE
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{fill_var}{A categorical variable to group and fill.}

\item{fill}{character. Color of all Density and Histograms. Overrided if fill_var is specified.}

\item{ncol}{integer. Number of facet columns.}

\item{bins}{integer. Number of bins. Overrided if show_histogram = FALSE.}

\item{alpha}{numeric. Opacity. Range: (0, 1).}

\item{scales}{character. Control facet x & y-axis ranges. Options include "fixed", "free", "free_y", "free_x".}

\item{show_density}{logical. whether to display density.}

\item{adjust}{numeric. Multiplicate bandwidth adjustment. Overrided if show_density = FALSE.}

\item{trans_x}{character. X-axis transformation.}

\item{trans_y}{character. X-axis transformation.}

\item{legend_position}{character. Legend position. Options include "right", "left", "top", "bottom", "none".}

\item{title}{character. Title for the plot.}

\item{x_lab}{character. X-axis label for the plot.}

\item{y_lab}{character. Y-axis label for the plot.}

\item{interactive}{logical. Whether or not to return an interactive plot (plotly).}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot.
}
\description{
Plot Histograms of all numeric variables from data frame.
}
\examples{


plotHistogram(iris)
plotHistogram(iris, fill = "steelblue")
plotHistogram(iris, alpha = .5, show_density = TRUE, adjust = 2)
plotHistogram(iris, fill = "steelblue", alpha = .7, show_density = TRUE, adjust = 2)
plotHistogram(iris, alpha = .8, fill_var = Species)
plotHistogram(iris, alpha = .8, fill_var = Species, legend_position = "bottom", ncol=2)
plotHistogram(iris, alpha = .4, fill_var = Species, show_density = TRUE, adjust = 2)

}
