% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly-themes.R
\name{plotly_dark}
\alias{plotly_dark}
\alias{plotly_white}
\title{Plotly Themes}
\usage{
plotly_dark(
  plot,
  config.scrollZoom = FALSE,
  config.displayModeBar = FALSE,
  config.responsive = TRUE,
  config.staticPlot = FALSE,
  layout.dragmode = TRUE,
  layout.autosize = TRUE,
  layout.width = NULL,
  font.color = "#ffffff",
  font.family = "Segoe UI",
  layout.background = "#292c30",
  ...
)

plotly_white(
  plot,
  config.scrollZoom = FALSE,
  config.displayModeBar = FALSE,
  config.responsive = TRUE,
  config.staticPlot = FALSE,
  layout.dragmode = TRUE,
  layout.autosize = TRUE,
  layout.width = NULL,
  font.color = "#1c1c1c",
  font.family = "Segoe UI",
  layout.background = "#ffffff",
  ...
)
}
\arguments{
\item{plot}{A Plotly chart object.}

\item{config.scrollZoom}{A logical value indicating whether to enable scroll-to-zoom functionality in the chart.}

\item{config.displayModeBar}{A logical value indicating whether to display the mode bar in the chart.}

\item{config.responsive}{A logical value indicating whether the chart should be responsive.}

\item{config.staticPlot}{A logical value indicating whether the chart should be static.}

\item{layout.dragmode}{A logical value indicating whether the chart should be draggable.}

\item{layout.autosize}{A logical value indicating whether the chart should automatically adjust its size.}

\item{layout.width}{The width of the chart in pixels.}

\item{font.color}{The font color for the chart.}

\item{font.family}{The font family for the chart.}

\item{layout.background}{The background color for the chart.}

\item{...}{Additional arguments to be passed to the Plotly chart object.}
}
\value{
A plotly object with a dark theme or white theme
}
\description{
Plotly Themes
}
\examples{


library(dplyr)
library(plotly)

iris \%>\%
  plot_ly() \%>\%
  add_trace(x = ~Sepal.Width, y = ~Sepal.Length, color = ~Species,
            type = "scatter", mode = "markers") \%>\%
  plotly_dark() \%>\%
  layout(
    title = list(text = "<b>Iris Sepal.Length vs Sepal.Width<b>"),
    legend=list(title=list(text='Species'))
  )

}
