% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNormality.R
\name{plotNormality}
\alias{plotNormality}
\title{Plot Normality of Numeric Variable}
\usage{
plotNormality(
  data,
  interactive = FALSE,
  color_density = "steelblue",
  color_qq = "steelblue",
  color_qq_line = "steelblue4",
  color_sqrt = "darkorange",
  color_log = "salmon"
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{interactive}{A boolean value that specifies whether the plot should be returned static (ggplot2) or interactive (Plotly).}

\item{color_density}{A string indicating the color to be used for the density plots. Default is "steelblue".}

\item{color_qq}{A string indicating the color to be used for the Q-Q plots. Default is "steelblue".}

\item{color_qq_line}{A string indicating the color to be used for the line in the Q-Q plots. Default is "steelblue4".}

\item{color_sqrt}{A string indicating the color to be used for the density plots of the sqrt-transformed variables. Default is "darkorange".}

\item{color_log}{A string indicating the color to be used for the density plots of the log-transformed variables. Default is "salmon".}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot
}
\description{
Plot Normality of Numeric Variable
}
\examples{

library(dplyr)

iris \%>\%
   select(Petal.Width) \%>\%
   plotNormality()

iris \%>\%
   select(Petal.Width, Sepal.Length) \%>\%
   plotNormality()

}
