% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScatter.R
\name{plotScatter}
\alias{plotScatter}
\title{Explore numeric columns with Scatterplot}
\usage{
plotScatter(
  data,
  y,
  fill_var = NULL,
  fill = NULL,
  ncol = 3,
  alpha = 1,
  size = 2,
  scales = "free",
  sample = NULL,
  smooth = FALSE,
  smooth_color = "dodgerblue4",
  smooth_fill = "gray",
  smooth_alpha = 0.3,
  smooth_method = "loess",
  smooth_se = FALSE,
  legend_position = "right",
  trans_x = "identity",
  trans_y = "identity",
  title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  interactive = FALSE
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{y}{A numeric column.}

\item{fill_var}{A categorical column to group and fill.}

\item{fill}{character. Color of the points. Overrided if fill_var is specified.}

\item{ncol}{integer. Number of facet columns.}

\item{alpha}{numeric. Opacity. Range: (0, 1).}

\item{size}{numeric. Size of the points.}

\item{scales}{character. Control facet x & y-axis ranges. Options include "fixed", "free", "free_y", "free_x".}

\item{sample}{integer. Number of rows to sample.}

\item{smooth}{logical. Whether or not to include a smooth line.}

\item{smooth_color}{character. Smooth line color}

\item{smooth_fill}{character. Confidence interval fill around the smooth line.}

\item{smooth_alpha}{numeric. Confidence interval opacity around the smooth line}

\item{smooth_method}{character. Smoothing method (function).}

\item{smooth_se}{logical. Whether or not to include a confidence interval around the smooth line.}

\item{legend_position}{character. Legend position. Options include "right", "left", "top", "bottom", "none".}

\item{trans_x}{character. X-axis transformation.}

\item{trans_y}{character. Y-axis transformation.}

\item{title}{character. Title for the plot.}

\item{x_lab}{character. X-axis label for the plot.}

\item{y_lab}{character. Y-axis label for the plot.}

\item{interactive}{logical. Whether or not to return an interactive plot (plotly).}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot.
}
\description{
Plot Scatterplot of all numeric variables from data frame fixed to a selected numeric variable.
}
\examples{

plotScatter(iris, y = Sepal.Length)

plotScatter(iris, y = Sepal.Length, fill_var = Species)

plotScatter(ggplot2::diamonds, y=price, sample = 1000, alpha = 0.4)

plotScatter(ggplot2::diamonds, y=price, sample = 1000, alpha = 0.4, fill="steelblue", smooth = TRUE)

plotScatter(ggplot2::diamonds, y=price, sample = 1000, alpha = 0.4, smooth = TRUE)


}
