% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotViolin.R
\name{plotViolin}
\alias{plotViolin}
\title{Explore numeric and categorical columns with Violin plot}
\usage{
plotViolin(
  data,
  target,
  fill = NULL,
  ncol = 3,
  scales = "free",
  alpha = 0.7,
  adjust = 2,
  trim = FALSE,
  title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  boxplot_show = FALSE,
  flip = FALSE
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{target}{A categorical column to group and fill}

\item{fill}{character. Color of all Violin. Overrided if fill_var is specified.}

\item{ncol}{integer. Number of facet columns.}

\item{scales}{character. Control facet x & y-axis ranges. Options include "fixed", "free", "free_y", "free_x".}

\item{alpha}{numeric. Opacity. Range: (0, 1).}

\item{adjust}{numeric. Multiplicate bandwidth adjustment.}

\item{trim}{A logical. Whether or not to trim tails.}

\item{title}{character. Title for the plot.}

\item{x_lab}{character. X-axis label for the plot.}

\item{y_lab}{character. Y-axis label for the plot.}

\item{boxplot_show}{logical. Whether or not to show boxplot.}

\item{flip}{logical. whether or not to flip cartesian coordinates.}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot.
}
\description{
Plot Violin of all variables fixing on a selected variable.
}
\examples{

plotViolin(iris, target = Species)

plotViolin(iris, target = Species, boxplot_show = TRUE)

plotViolin(iris, target=Species, boxplot_show=TRUE, flip=TRUE)

plotViolin(iris, target=Species, fill="steelblue")

mtcars <- mtcars \%>\% dplyr::as_tibble() \%>\%
  dplyr::mutate(dplyr::across(c(am, carb, cyl, gear, vs), as.factor))

plotViolin(mtcars, target = mpg)

plotViolin(mtcars, target = mpg, boxplot_show=TRUE)

plotViolin(mtcars, target = mpg, fill = "steelblue")

}
