% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincloud_plot.R
\name{raincloud_plot}
\alias{raincloud_plot}
\title{Plot Raincloud Plot}
\usage{
raincloud_plot(
  data,
  x,
  fill_var = NULL,
  fill = NULL,
  alpha = 0.8,
  box_alpha = 0.5,
  adjust = 1,
  jitter_sample = 100,
  jitter_alpha = 0.5,
  jitter_height = 0.2,
  jitter_width = 0.2,
  scales = "free",
  ncol = 1,
  title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  interactive = FALSE
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{x}{A numerical column.}

\item{fill_var}{A categorical column to group and fill.}

\item{fill}{character. Color of the points. Overrided if fill_var is specified.}

\item{alpha}{numeric. Opacity of density curve. Range: (0, 1).}

\item{box_alpha}{numeric. Opacity of box plot.}

\item{adjust}{numeric. Multiplicate bandwidth adjustment..}

\item{jitter_sample}{integer. Number of rows to sample.}

\item{jitter_alpha}{numeric. Opacity. Range: (0, 1).}

\item{jitter_height}{numeric. Height.}

\item{jitter_width}{numeric. Width.}

\item{scales}{character. Control facet x & y-axis ranges. Options include "fixed", "free", "free_y", "free_x".}

\item{ncol}{integer. Number of facet columns.}

\item{title}{character. Title for the plot.}

\item{x_lab}{character. X-axis label for the plot.}

\item{y_lab}{character. Y-axis label for the plot.}

\item{interactive}{logical. Whether or not to return an interactive plot (plotly).}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot.
}
\description{
Plot Raincloud Plot in ggplot2
}
\examples{

raincloud_plot(iris, x = Sepal.Length)

raincloud_plot(iris, x = Sepal.Length, fill = "orange")

raincloud_plot(iris, x = Sepal.Length, fill_var = Species, ncol = 1, fill = "steelblue")

raincloud_plot(ggplot2::diamonds, x = carat, fill_var = cut, ncol = 2, adjust = 2)


}
