% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_chart.R
\name{scatter_chart}
\alias{scatter_chart}
\title{Scatter Chart}
\usage{
scatter_chart(
  data,
  x,
  y,
  fill_var = NULL,
  facet_var = NULL,
  color = "#4E79A7",
  alpha = 1,
  size = 1.5,
  ncol = 2,
  scales = "fixed",
  highlight = NULL,
  compare = FALSE,
  trans_x = "identity",
  trans_y = "identity",
  smooth = FALSE,
  method = "lm",
  se = FALSE,
  title = NULL,
  x_lab = NULL,
  y_lab = NULL,
  legend_position = "right",
  interactive = FALSE
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{x}{A numeric column.}

\item{y}{A numeric column.}

\item{fill_var}{A numeric or categorical variable to group and fill.}

\item{facet_var}{A categorical variable used to create facet panels.}

\item{color}{character. Color of the points. Overrided if fill_var is specified.}

\item{alpha}{numeric. Opacity. Range: (0, 1).}

\item{size}{numeric. Size of the points.}

\item{ncol}{integer. Number of facet columns.}

\item{scales}{character. control facet x & y-axis ranges. Options include "fixed", "free", "free_y", "free_x".}

\item{highlight}{A vector of categories to highlight.}

\item{compare}{logical. Whether or not to compare a group with the others.}

\item{trans_x}{character. X-axis transformation.}

\item{trans_y}{character. Y-axis transformation.}

\item{smooth}{logical. Whether or not to include a smooth line.}

\item{method}{character. Smoothing method (function).}

\item{se}{logical. Whether or not to include a confidence interval around the smooth line.}

\item{title}{character. Title for the plot.}

\item{x_lab}{character. X-axis label for the plot.}

\item{y_lab}{character. Y-axis label for the plot.}

\item{legend_position}{character. Legend position. Options include "right", "left", "top", "bottom", "none".}

\item{interactive}{logical. Whether or not to return an interactive plot (plotly).}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot.
}
\description{
Plot Scatter Chart in ggplot2
}
\examples{
scatter_chart(iris, x = Petal.Width, y=Sepal.Length)

scatter_chart(iris, x = Petal.Width, y=Sepal.Length, facet_var = Species)

scatter_chart(iris, x = Petal.Width, y=Sepal.Length, facet_var = Species,
              compare = TRUE)

scatter_chart(iris, x = Petal.Width, y=Sepal.Length, fill_var = Species)

scatter_chart(iris, x = Petal.Width, y=Sepal.Length, fill_var = Species,
              highlight = c("setosa"))

scatter_chart(iris, x = Petal.Width, y=Sepal.Length, fill_var = Species,
              smooth = TRUE, se = TRUE)

scatter_chart(iris, x = Petal.Width, y=Sepal.Length, fill_var = Species,
              facet_var = Species)

scatter_chart(iris, x = Petal.Width, y=Sepal.Length, fill_var = Species,
              facet_var = Species, compare = TRUE)

scatter_chart(iris, x = Petal.Width, y=Sepal.Length, fill_var = Species,
              facet_var = Species, compare = TRUE, highlight = "setosa")

scatter_chart(iris, x = Petal.Width, y=Sepal.Length, fill_var = Species,
              facet_var = Species, scales = "free")


}
