% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yoy_barchart.R
\name{yoy_barchart}
\alias{yoy_barchart}
\title{YoY Chart by Category with sparklines}
\usage{
yoy_barchart(
  data,
  date,
  category,
  value,
  current_year,
  previous_year,
  summarise_by = "month",
  top_n = 10,
  show_label = TRUE,
  label_color = "grey30",
  label_size = 3,
  title = NULL,
  title_size = 14,
  title_hjust = 0.5,
  palette_gradient = "Green",
  bar_width = 0.5,
  linepy_color = "grey20",
  linepy_width = 1,
  expand_mult = NULL,
  axistext_size = 9,
  ts_linecolor = "steelblue",
  ts_linewidth = 0.5,
  ts_pointsize = 0.85,
  interactive = FALSE
)
}
\arguments{
\item{data}{A data-frame or tibble.}

\item{date}{A date or date time variable.}

\item{category}{A categorical variable.}

\item{value}{A numerical variable.}

\item{current_year}{integer. Current year.}

\item{previous_year}{integer. Prior year.}

\item{summarise_by}{character. The time unit to summarise.}

\item{top_n}{integer. Top n levels.}

\item{show_label}{logical. Whether or not to include label.}

\item{label_color}{numeric. Label color.}

\item{label_size}{numeric. Label size.}

\item{title}{character. Title for the plot.}

\item{title_size}{numeric. The size of the title.}

\item{title_hjust}{numeric. Horizontal justification of the title.}

\item{palette_gradient}{numeric. Divergent color palette. Color palettes used in Tableau (ggthemes).}

\item{bar_width}{numeric. Bar Width.}

\item{linepy_color}{character. Color of the prior year line in the bar chart.}

\item{linepy_width}{numeric. Size of the prior year line in the bar chart.}

\item{expand_mult}{A numeric vector of length 2 indicating the expansion multiplier for the y-axis in bar chart.}

\item{axistext_size}{numeric. Size of the y-axis text in bar chart.}

\item{ts_linecolor}{character. Color of the lines in line chart.}

\item{ts_linewidth}{numeric. Width of the lines in line chart.}

\item{ts_pointsize}{numeric. Size of the points in line chart.}

\item{interactive}{logical. Whether or not to return an interactive plot (plotly).}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot.
}
\description{
Plot a YoY chart by Category with sparklines in ggplot2.
}
\examples{

data <- readr::read_csv(
  file = "https://raw.githubusercontent.com/zpio/datasets/main/sample_superstore.csv"
)

yoy_barchart(
  data = data,
  date = order_date,
  category = state,
  value = sales,
  current_year = 2021,
  previous_year = 2020,
  title = "Sales by Category 2021 (YoY \%)"
)

}
