% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yoy_barchart0.R
\name{yoy_barchart0}
\alias{yoy_barchart0}
\title{YoY Chart by Category}
\usage{
yoy_barchart0(
  data,
  category,
  cy_var,
  py_var,
  label_var_p1 = NULL,
  label_var_p2 = NULL,
  label_color = "grey30",
  label_size = 3,
  title = NULL,
  title_size = 14,
  title_hjust = 0.5,
  palette_gradient = "Green",
  width_bar = 0.5,
  expandmult_p1 = c(0, 0.4),
  expandmult_p2 = c(0.5, 0.5),
  axis_text_size = 9
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{category}{A categorical variable that represents the y-axis.}

\item{cy_var}{integer. Current year.}

\item{py_var}{integer. Prior year.}

\item{label_var_p1}{A categorical variable to label the bar chart.}

\item{label_var_p2}{A categorical variable to label the YoY\% plot.}

\item{label_color}{character. Label color.}

\item{label_size}{numeric. Label size.}

\item{title}{character. Title for the plot.}

\item{title_size}{numeric. Title size.}

\item{title_hjust}{numeric. Horizontal justification of the title.}

\item{palette_gradient}{numeric. Divergent color palette. Color palettes used in Tableau (ggthemes).}

\item{width_bar}{numeric. Width of the bars.}

\item{expandmult_p1}{A numeric vector of length 2 indicating the expansion multiplier for the x-axis in bar chart.}

\item{expandmult_p2}{A numeric vector of length 2 indicating the expansion multiplier for the x-axis in YoY\% plot.}

\item{axis_text_size}{logical. Whether or not to return an interactive plot (plotly).}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot.
}
\description{
Plot a YoY Chart by Category in ggplot2.
}
\examples{
data_sales <- data.frame(
  category = c("Furniture","Office Supplies","Technology"),
  sales = c(414288.7, 430037.2, 498095.0),
  sales_2021 = c(215387.3, 246097.2, 271730.8),
  sales_2020 = c(198901.4, 183940.0, 226364.2)
)

yoy_barchart0(
  data = data_sales,
  category = category,
  cy_var = sales_2021,
  py_var = sales_2020
)

}
