% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yoy_category.R
\name{yoy_category}
\alias{yoy_category}
\title{Compute year-over-year growth by category}
\usage{
yoy_category(data, date, category, value, current_year, previous_year)
}
\arguments{
\item{data}{A data-frame or tibble.}

\item{date}{A date variable.}

\item{category}{A categorical variable.}

\item{value}{A numerical variable.}

\item{current_year}{numeric. Current year.}

\item{previous_year}{numeric. Prior year.}
}
\value{
A tibble
}
\description{
This function calculates the year-over-year (YoY) growth of a given numerical variable, grouped by a category variable. It filters the data by the specified date variable, selecting only the years specified by current_year and previous_year.
}
\examples{
library(readr)
library(dplyr)
data <- read_csv(
  file = "https://raw.githubusercontent.com/zpio/datasets/main/sample_superstore.csv"
)

data_yoy <-
  yoy_category(
    data = data,
    date = order_date,
    category = category,
    value = sales,
    current_year = 2021,
    previous_year = 2020
  )

}
