% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yoy_chart.R
\name{yoy_chart}
\alias{yoy_chart}
\title{Year-over-Year chart}
\usage{
yoy_chart(
  data,
  date,
  value,
  prior_year,
  current_year,
  summarise_by = "month",
  color_prior = "gray",
  color_current = "#4E79A7",
  linewidth_prior = 1,
  linewidth_current = 1,
  label_size = 3.5,
  expand_mult_x = c(0.04, 0.12),
  title = NULL,
  y_lab_p1 = NULL,
  y_lab_p2 = NULL,
  color_positive = "green3",
  color_negative = "red3",
  label_size_p2 = 3,
  label_color_p2 = "gray10",
  expand_mult_y_p2 = c(0.25, 0.25),
  pct = FALSE
)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{date}{A date or date time variable.}

\item{value}{A numeric variable.}

\item{prior_year}{integer. Prior year.}

\item{current_year}{integer. Current year.}

\item{summarise_by}{character. Time unit to summarise the data.}

\item{color_prior}{character. Line color of the prior year.}

\item{color_current}{character. Line color of the current year.}

\item{linewidth_prior}{numeric. Line width of the prior year.}

\item{linewidth_current}{numeric. Line width for the current year.}

\item{label_size}{numeric. Label font size for the current year.}

\item{expand_mult_x}{A numeric vector of length 2 indicating the expansion of the x-axis in the line chart and bar chart.}

\item{title}{character. Title for the plot.}

\item{y_lab_p1}{character. Y-axis label for the line chart.}

\item{y_lab_p2}{character. Y-axis label for the bar chart.}

\item{color_positive}{character. Color for positive percentage changes.}

\item{color_negative}{character. Color for negative percentage changes.}

\item{label_size_p2}{numeric. Label font size for the bar chart.}

\item{label_color_p2}{character. Color labels on the bar chart.}

\item{expand_mult_y_p2}{A numeric vector of length 2 indicating the amount to expand the y-axis limits for the bar chart.}

\item{pct}{logical. Whether or not to format the values as percentages.}
}
\value{
A ggplot object
}
\description{
Year-over-Year chart in ggplot2.
}
\examples{
library(readr)
library(dplyr)
data <- read_csv(
  file = "https://raw.githubusercontent.com/zpio/datasets/main/sample_superstore.csv"
)

yoy_chart(
  data = data,
  date = order_date,
  value = sales,
  prior_year = 2020,
  current_year = 2021
)
}
